[S11P31A307-262] Fix: 기능 테스트 진행하며 오류 수정

+++ shieldrone-main-server/src/main/java/com/ssafy/shieldron/repository/DroneRepository.java
@@ -19,5 +19,5 @@ public interface DroneRepository extends JpaRepository<Drone, Integer> {
     AND d.battery >= :requiredBattery
     ORDER BY d.battery DESC
 """)
-    Optional<Drone> findFirstAvailableDroneWithLock(@Param("requiredBattery") double requiredBattery);
+    Optional<Drone> findFirstAvailableDroneWithLock(@Param("requiredBattery") Integer requiredBattery);
 }
